requirements.txt
fastapi==0.111.1
langchainhub==0.1.20
llama-index==0.10.57
llama-index-core==0.10.57
llama-index-embeddings-azure-openai==0.1.11
llama-index-embeddings-text-embeddings-inference==0.1.4
llama-index-llms-azure-openai==0.1.10
llama-index-llms-llama-cpp==0.1.4
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-azure-openai==0.1.4
llama-index-postprocessor-rankgpt-rerank==0.1.6
llama-index-readers-file==0.1.30
llama-index-readers-web==0.1.23
nebula3-python==3.8.1
neo4j==5.22.0
pika==1.3.2
pillow==10.4.0
poppler-utils==0.1.0
pydub==0.25.1
PyMuPDF==1.24.8
pypdf==4.3.1
pytest-playwright==0.5.1
python-magic==0.4.27
qdrant-client==1.10.1
ragas==0.1.11
scikit-learn==1.5.1
seaborn==0.13.2
trulens-eval==0.33.0
uvicorn[standard]==0.30.3
wikipedia==1.4.0
duckdb==1.0.0
openai-whisper==20231117